Refactor code and clean up for review
